# Vectorize `tsince` in `propagation.sgp4()`