--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\Sholes\\AppData\\Local\\Continuum\\miniconda3\\envs\\pydata\\python.exe"
+}