--- a/setup.py
+++ b/setup.py
@@ -1,5 +1,4 @@
-import sys
-from distutils.core import setup, Extension
+from distutils.core import setup
 from textwrap import dedent
 
 import sgp4, sgp4.model
@@ -7,18 +6,6 @@ import sgp4, sgp4.model
 description, long_description = sgp4.__doc__.split('\n', 1)
 satdoc = dedent(sgp4.model.Satellite.__doc__.split('\n', 1)[1])
 long_description = long_description.replace('entry.', 'entry.' + satdoc)
-ext_modules = []
-
-if sys.version_info[0] == 3:
-    ext_modules.append(Extension(
-        'sgp4.vallado_cpp',
-        sources = [
-            'extension/SGP4.cpp',
-            'extension/wrapper.cpp',
-        ],
-        # extra_compile_args=['-fopenmp'],
-        # extra_link_args=['-fopenmp'],
-    ))
 
 setup(name = 'sgp4',
       version = sgp4.__version__,
@@ -39,11 +26,7 @@ setup(name = 'sgp4',
         'Programming Language :: Python :: 3.2',
         'Programming Language :: Python :: 3.3',
         'Programming Language :: Python :: 3.4',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
         'Topic :: Scientific/Engineering :: Astronomy',
         ],
       packages = ['sgp4'],
-      ext_modules = ext_modules,
-)
+      )