@dsholes 2019-04-15
I just noticed an issue. Using `@jit` when `tsince` is a numpy array is very slow the first time `sgp4` is called. I think it's related to this [stackoverflow post](https://stackoverflow.com/questions/50658884/why-this-numba-code-is-6x-slower-than-numpy-code).

In short, it seems like creating a lot of numpy arrays within a numba function is not a good idea. However, numpy outperforms a numba function by an order of magnitude for the case of sgp4 looping over a large `tsince` array.

Thus, for a vectorized version using numpy arrays, it is best to avoid using `@jit`. I'll update the pull request.