@brandon-rhodes 2019-08-05
Alas, that I still haven't spent time with this pull request from April! But my schedule now lets me jump in.

My big question about turning this into vectors is the same question I see that you've asked in a comment: what happens with an `if` statement that adjusts parameters, but should only do so for satellites in particular circumstances? I think the PR would need to become more sophisticated in those circumstances, doing tricks like I learned to do in Skyfield where a true/false bitmask applies an adjustment to some satellites but not others.

First: what this PR really needs is a comprehensive test. The test needs to get all the satellites in the test suite data set, from the satellite file that came with SGP4; it needs to turn them into a single big vector of all those satellites; and it needs to send them all at once through SGP4. If they all come out with the positions expected, then we'll have confidence that the vector routines are behaving.

Could you add such a test to the PR and report back about how many satellites currently give good positions when all are submitted at the same time as a vector? Thanks!