@dsholes 2019-08-07
The "satellite file" you're referring to is ["SGP4-VER.TLE"](https://github.com/brandon-rhodes/python-sgp4/blob/master/sgp4/SGP4-VER.TLE), correct?

I'll work on the test you're describing. 

The one difference is I won't send all satellites through as a big vector. Based on the way `sgp4` and the `skyfield` earth sat `.at()` methods currently function (treating satellites as objects), I think it makes more sense to continue running one satellite at a time (for-loop method).  The main difference between the normal code and this PR is that the **time array** can be input as a vector, not satellite objects.

But I'll run all TLEs (satellites) in ["SGP4-VER.TLE"](https://github.com/brandon-rhodes/python-sgp4/blob/master/sgp4/SGP4-VER.TLE), and use the same large time vector for all. I'll then check agreement with the normal sgp4 positions, which uses a for loop to move through singular times.

Should the time array include times before the epoch of each satellite (negative tsince)?

Also regarding this:
> where a true/false bitmask applies an adjustment to some satellites but not others.

I've been using `np.where(bool_mask, value_if_true, value_if_false)` whenever I encounter an if-else statement that becomes vectorized with `t`. I imagine that's what you're referring to.