@brandon-rhodes 2019-12-02
Did you ever get results from testing your vectorized algorithm against SGP4-VER.TLE? In the meantime, I should note that conversations have advanced about whether this package, or another one, should in addition to this Python SGP4 implementation, wrap the official C implementation to provide extra speed for users needing lots of satellites. So, if it turns out your improvements here struggle to correctly propagate lots of satellites, we maybe could use the C version as an alternative way of producing some speedup for users who need it.