@dsholes 2019-12-05
Thanks for the reminder, I got back to this yesterday. The short answer is, the vectorization is working for `satrec.method == 'n'` (near earth), but it would take a big effort and massive changes to the code to get it working for `satrec.method == 'd'` (deep space).

For now, if it's `'d'` I'd stick to the for-loop, and for `'n'` I'd use the vectorized code. I use SGP4 exclusively for LEO sats so this wouldn't be a deal breaker for me, but from a maintainability standpoint, I think it would be much easier / cleaner to go the Cython route. I'm very curious how the speed will compare for near earth cases.

I'm not very familiar with Cython, so sorry if this doesn't make any sense, but would the for-loop for a large time array be "wrapped" so that it's implemented in C instead of in Python? I guess what I'm asking is, if I pass a large time array to sgp4 or skyfield, will the for-loop still be run in Python, and only the calculations take place in the C code, or will everything take place "inside" the C code?

Also, I'm struggling to integrate everything into the current `sgp4.tests` (it's difficult since the tests were written for single "time" calls not, arrays of time). In case you're interested in the vectorized code and my attempt at integrating it into your existing tests, I can push my changes thus far, but fair warning, it's a bit of a mess.