@brandon-rhodes 2019-12-05
If the vectorized version only works for one class of satellites, then my guess, alas, is that it would be too confusing for users to merge it into the main package — after all, satellites do not come labelled in the TLE file with whether they are deep space or not; it's the SGP4 import routine that figures that out based on the orbital parameters.

Over the past couple of days I have tried making a minimal wrapper around the raw C++ code from the maintainer for SGP4 itself, and its performance is promising. A couple of years ago I would not have had any confidence in getting a C-language module pushed to users, but today there are much better mechanisms for building Windows and Mac (and Linux) packages using CI services and then getting them all uploaded to PyPI.

Anyway, thanks VERY much for the work you did here to see if the algorithm, without a huge rewrite, would be susceptible to processing vectors of input dates. It's unfortunate that the answer looks to be "no", but that gives us very useful signal about where to take the project next!

Should we close this pull request, or are there some last details you'd like to discuss first?