@dsholes 2019-12-06
In my latest commit I added [new tests](https://github.com/brandon-rhodes/python-sgp4/blob/809ca0feffcdd4e516e530ce5cbe6ad13d455f07/test_ds_main.ipynb), in case you're ever interested in integrating these kinds of changes in the future.

Just to clarify, I hide the logic for `'d'` vs `'n'` and single `tsince` vs array of `tsince` within `propagation.py`, so it would never fall on the user to make that decision. 

However, I still agree that I think it makes a lot more sense to wrap the C++ code. You may want to take a look at the [tests](https://github.com/brandon-rhodes/python-sgp4/blob/809ca0feffcdd4e516e530ce5cbe6ad13d455f07/test_ds_main.ipynb) (plots at the bottom of the notebook) to see the performance difference for large time arrays. I would be interested in running a similar test in the future to see how the Cython performance gain compares to the vectorized (numpy) code.

So yes, feel free to close this pull request, and once again thank you so much for all of your work on this library and skyfield. They're both amazing.