@brandon-rhodes 2019-12-09
> So yes, feel free to close this pull request, and once again thank you so much for all of your work on this library and skyfield. They're both amazing.

Thanks, it's very encouraging to hear when the libraries have helped folks with their projects!